ipc-justice-aid-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
django-extensions==3.2.3
django-ratelimit==4.1.0
cryptography==41.0.7
pyjwt==2.8.0
Pillow==10.1.0

# Production and Azure specific packages
gunicorn==21.2.0
whitenoise==6.6.0
django-storages[azure]==1.14.2
azure-storage-blob==12.19.0
sentry-sdk[django]==1.38.0

# Development and testing
pytest-django>=4.5.0
pytest>=7.0.0
factory-boy>=3.2.0
coverage>=7.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3